# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API",
    "description": "API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3010",
      "name": "DEV"
    }
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by id",
        "description": "Get a user by id",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update a user",
        "description": "Update a user",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "description": "User object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user",
        "description": "Delete a user",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/api/v1/requests": {
      "get": {
        "tags": ["Requests"],
        "summary": "Get all requests",
        "description": "Get all requests",
        "operationId": "getRequests",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": ["Requests"],
        "summary": "Create a request",
        "description": "Create a request",
        "operationId": "createRequest",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/requests/{id}": {
      "patch": {
        "tags": ["Requests"],
        "summary": "Update a request",
        "description": "Update a request",
        "operationId": "updateRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of request to return",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": ["Requests"],
        "summary": "Delete a request",
        "description": "Delete a request",
        "operationId": "deleteRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of request to return",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": ["Requests"],
        "summary": "Approve/Reject a request",
        "description": "Accept a request",
        "operationId": "acceptRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of request to return",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/changeRequestStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accepted"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "department_name": {
          "type": "string"
        }
      }
    },
...
    "Request": {
      "type": "object",
      "properties": {
        "startDate": {
          "type": "string",
          "format": "date",
          "default": "2022-05-01"
        },
        "numberOfDays": {
          "type": "integer",
          "default": 1
        },
        "type": {
          "type": "string",
          "default": "Vacation"
        },
        "reason": {
          "type": "string"
        }
      }
    },
    "changeRequestStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "default": "pending"
        }
      }
    },
...
  }
}